feat: add 2024-2025 graph + breadcrumbs
