+++ mauritius_tax_calculator/src/pages/2024-2025/components/YearlyCalculations.tsx
@@ -15,22 +15,11 @@ import { useFormikContext } from "formik";
 import { Trans, useTranslation } from "next-i18next";
 import { useEffect, useState } from "react";
 import {
-  CSG_DECREASED_RATE,
-  CSG_INCREASED_RATE,
-  CSG_MAX_MONTHLY_DOMESTIC_LIMIT,
-  CSG_MAX_YEARLY_DOMESTIC_LIMIT,
-  CSG_MONTHLY_BASE_SALARY_LIMIT,
-  CSG_YEARLY_BASE_SALARY_LIMIT,
   CURRENT_FINANCIAL_YEAR_NAMESPACE,
-  INITIAL_YEARLY_TAXABLE_BRACKETS,
   MRA_YEARLY_MAX_NON_TAXABLE_TRAVELING_ALLOWANCE,
-  NSF_MAX_YEARLY_INSURABLE_BASIC_WAGE,
-  NSF_MIN_YEARLY_INSURABLE_BASIC_WAGE_HOUSEHOLD_EMPLOYEE,
-  NSF_MIN_YEARLY_INSURABLE_BASIC_WAGE_NORMAL_EMPLOYEE,
-  NSF_RATE,
-  YEARLY_IET_DEPENDENT_DEDUCTIONS,
 } from "../reusables";
 import { TaxCalcRow, YearlyFormValues } from "../types";
+import { retrieveYearlyValues } from "../graph/components/reusables";
 
 export default function YearlyCalculations() {
   const { t: tCommon } = useTranslation("common");
@@ -72,137 +61,33 @@ export default function YearlyCalculations() {
   useEffect(() => {
     if (isValid) {
       try {
-        const newMonthlyBaseSalary = new Decimal(values.monthlyBaseSalary);
-        const newTotalBaseSalary = new Decimal(values.totalBaseSalary);
-        const newEOYBonus = new Decimal(values.eoyBonus);
-        const newTravelingAllowance = new Decimal(values.travelingAllowance);
-        const totalBaseSalaryTimes25Percent = newTotalBaseSalary.dividedBy(4);
-        const newMaxNonTaxableTravelingAllowance = Decimal.min(
-          MRA_YEARLY_MAX_NON_TAXABLE_TRAVELING_ALLOWANCE,
-          totalBaseSalaryTimes25Percent
-        );
-        const newTaxableTravelingAllowance =
-          newTravelingAllowance.lessThanOrEqualTo(
-            newMaxNonTaxableTravelingAllowance
-          )
-            ? new Decimal(0)
-            : newTravelingAllowance.sub(newMaxNonTaxableTravelingAllowance);
-        const newPerformanceBonus = new Decimal(values.performanceBonus);
-        const newOtherTaxableIncome = new Decimal(values.otherTaxableIncome);
-        let newChargeableIncome = newTotalBaseSalary
-          .add(newEOYBonus)
-          .add(newTaxableTravelingAllowance)
-          .add(newPerformanceBonus)
-          .add(newOtherTaxableIncome);
-        let newTotalIncome = newTotalBaseSalary
-          .add(newEOYBonus)
-          .add(newTravelingAllowance)
-          .add(newPerformanceBonus)
-          .add(newOtherTaxableIncome);
-        const newIETDeductions =
-          YEARLY_IET_DEPENDENT_DEDUCTIONS[values.numOfDependents];
-        const newOtherTaxDeductions = new Decimal(values.otherTaxDeductions);
-        const newTotalDeductions = newIETDeductions.add(newOtherTaxDeductions);
-        newChargeableIncome = newChargeableIncome.sub(newTotalDeductions);
-        let remainder = newChargeableIncome;
-        let newPAYE = new Decimal(0);
-        const newYearlyPAYECalcRows = [
-          ...INITIAL_YEARLY_TAXABLE_BRACKETS.map((bracket, index) => {
-            if (!bracket.limit || remainder.lessThan(bracket.limit)) {
-              const taxableAmount = remainder;
-              const taxCharged = taxableAmount.mul(bracket.rate);
-              newPAYE = newPAYE.add(taxCharged);
-              remainder = remainder.sub(remainder);
-
-              return {
-                key: String(index),
-                taxableLimit: bracket.limit
-                  ? decimalToString(bracket.limit)
-                  : tCurrentYear(
-                      "year.output.paye.table.taxableLimits.remainder"
-                    ),
-                taxableAmount: decimalToString(taxableAmount, 2),
-                taxRate: decimalToString(bracket.rate.mul(100)),
-                taxCharged: decimalToString(taxCharged, 2),
-              };
-            }
-
-            const taxableAmount = bracket.limit;
-            const taxCharged = taxableAmount.mul(bracket.rate);
-            newPAYE = newPAYE.add(taxCharged);
-            remainder = remainder.sub(bracket.limit);
-
-            return {
-              key: String(index),
-              taxableLimit: decimalToString(bracket.limit),
-              taxableAmount: decimalToString(taxableAmount, 2),
-              taxRate: decimalToString(bracket.rate.mul(100)),
-              taxCharged: decimalToString(taxCharged, 2),
-            };
-          }),
-          {
-            key: "aggregations",
-            taxableLimit: tCurrentYear(
-              "year.output.paye.table.taxableLimits.taxCharged"
-            ),
-            taxableAmount: decimalToString(newChargeableIncome, 2),
-            taxRate: null,
-            taxCharged: decimalToString(newPAYE, 2),
-          },
-        ];
-        const isExemptFromCSG =
-          (!values.isCitizen && !values.isResident) ||
-          (values.isPublicSector && !values.isPRB);
-        const isExemptFromCSGOnTotalBaseSalary =
-          isExemptFromCSG ||
-          (values.isInDomesticService &&
-            newTotalBaseSalary.lessThan(CSG_MAX_YEARLY_DOMESTIC_LIMIT));
-        const isExemptFromCSGOnMonthlyBaseSalary =
-          isExemptFromCSG ||
-          (values.isInDomesticService &&
-            newMonthlyBaseSalary.lessThan(CSG_MAX_MONTHLY_DOMESTIC_LIMIT));
-        let newTotalBaseSalaryCSGRate = new Decimal(0);
-        let newEOYBonusCSGRate = new Decimal(0);
-        const isExemptFromNSF = ["under18", "70AndOver"].includes(values.age);
-        const newNSFRate = isExemptFromNSF ? new Decimal(0) : NSF_RATE;
-        const minNSFInsurableSalary = values.isInDomesticService
-          ? NSF_MIN_YEARLY_INSURABLE_BASIC_WAGE_HOUSEHOLD_EMPLOYEE
-          : NSF_MIN_YEARLY_INSURABLE_BASIC_WAGE_NORMAL_EMPLOYEE;
-        const maxNSFInsurableSalary = NSF_MAX_YEARLY_INSURABLE_BASIC_WAGE;
-        const newNSFInsurableSalary = new Decimal(
-          newTotalBaseSalary.lessThan(minNSFInsurableSalary)
-            ? 0
-            : newTotalBaseSalary.greaterThan(maxNSFInsurableSalary)
-              ? maxNSFInsurableSalary
-              : newTotalBaseSalary
-        );
-
-        if (!isExemptFromCSGOnTotalBaseSalary) {
-          newTotalBaseSalaryCSGRate = newTotalBaseSalary.lessThanOrEqualTo(
-            CSG_YEARLY_BASE_SALARY_LIMIT
-          )
-            ? CSG_DECREASED_RATE
-            : CSG_INCREASED_RATE;
-        }
-
-        if (!isExemptFromCSGOnMonthlyBaseSalary) {
-          newEOYBonusCSGRate = newMonthlyBaseSalary.lessThanOrEqualTo(
-            CSG_MONTHLY_BASE_SALARY_LIMIT
-          )
-            ? CSG_DECREASED_RATE
-            : CSG_INCREASED_RATE;
-        }
-
-        const newTotalBaseSalaryCSG = newTotalBaseSalary.mul(
-          newTotalBaseSalaryCSGRate
-        );
-        const newEOYBonusCSG = newMonthlyBaseSalary.mul(newEOYBonusCSGRate);
-        const newCSG = newTotalBaseSalaryCSG.add(newEOYBonusCSG);
-        const newNSF = isExemptFromNSF
-          ? new Decimal(0)
-          : newNSFInsurableSalary.mul(newNSFRate);
-        const newTotalTaxes = newPAYE.add(newCSG).add(newNSF);
-        const newIncomeAfterTaxes = newTotalIncome.sub(newTotalTaxes);
+        const {
+          newMonthlyBaseSalary,
+          newTotalBaseSalary,
+          newEOYBonus,
+          newTravelingAllowance,
+          newMaxNonTaxableTravelingAllowance,
+          newTaxableTravelingAllowance,
+          newPerformanceBonus,
+          newOtherTaxableIncome,
+          newTotalIncome,
+          newIETDeductions,
+          newOtherTaxDeductions,
+          newTotalDeductions,
+          newChargeableIncome,
+          newPAYE,
+          newTotalBaseSalaryCSGRate,
+          newEOYBonusCSGRate,
+          newTotalBaseSalaryCSG,
+          newEOYBonusCSG,
+          newCSG,
+          newNSFRate,
+          newNSFInsurableSalary,
+          newNSF,
+          newTotalTaxes,
+          newYearlyPAYECalcRows,
+          newIncomeAfterTaxes,
+        } = retrieveYearlyValues({ values, tCurrentYear });
 
         setMonthlyBaseSalary(newMonthlyBaseSalary);
         setTotalBaseSalary(newTotalBaseSalary);
